refactor: Update online component of new initialize script
This references https://github.com/JWWeatherman/yeticold/pull/74

In order for the above linked PR to work, the online code must be
adjusted to make it visible to the user.

Also, apt-get has been changed to apt which is better for user facing
output and is otherwise identical.

Usage of semi-colons between multiple commands has been changed to AND
operators &&. This waits for the previous command to succeed before
continuing to the second command and is better practice.